feat/3: style - set up tailwind css
