--- apps/web/src/components/common/layout/layout.tsx
@@ -1,17 +1,0 @@
-import Head from "next/head";
-
-export default function Layout({
-  children,
-}: {
-  children: React.ReactNode;
-}): JSX.Element {
-  return (
-    <>
-      <Head>
-        <title>Scrapper</title>
-        <meta name="description" content="Scrapper Web" />
-      </Head>
-      <main className={"container"}>{children}</main>
-    </>
-  );
-}
