+++ apps/web/src/pages/_app.tsx
@@ -1,14 +1,21 @@
 import type { AppProps } from "next/app";
+import Head from "next/head";
 
-import Layout from "../components/common/layout";
+import PageLayout from "../components/layout";
 
-import "../styles/reset.css";
 import "../styles/globals.css";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
+    <>
+      <Head>
+        <title>Scrapper</title>
+        <meta name="description" content="Scrapper Web" />
+        <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
+      </Head>
+      <PageLayout>
+        <Component {...pageProps} />
+      </PageLayout>
+    </>
   );
 }
