+++ apps/web/src/pages/index.tsx
@@ -1,11 +1,3 @@
-export const metadata = {
-  title: "Scrapper",
-};
-
 export default function ScrapperHome(): JSX.Element {
-  return (
-    <h1 className="title">
-      <span>Scrapper</span>
-    </h1>
-  );
+  return <h1>Scrapper</h1>;
 }
